requirements.txt
streamlit==0.83.0
pandas==1.2.5
pillow==8.2.0
base58==2.1.0
biopython=1.79